fix roundrobin policy
