+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/sendsemantics/datatransferpolicy/HashBasedShufflePolicy.scala
@@ -12,50 +12,10 @@ class HashBasedShufflePolicy(
     batchSize: Int,
     val hashFunc: ITuple => Int,
     receivers: Array[ActorVirtualIdentity]
-) extends DataSendingPolicy(policyTag, batchSize, receivers) {
-  var batches: Array[Array[ITuple]] = _
-  var currentSizes: Array[Int] = _
+) extends ParallelBatchingPolicy(policyTag, batchSize, receivers) {
 
-  initializeInternalState(receivers)
-
-  override def noMore(): Array[(ActorVirtualIdentity, DataPayload)] = {
-    val receiversAndBatches = new ArrayBuffer[(ActorVirtualIdentity, DataPayload)]
-    for (k <- receivers.indices) {
-      if (currentSizes(k) > 0) {
-        receiversAndBatches.append(
-          (receivers(k), DataFrame(batches(k).slice(0, currentSizes(k))))
-        )
-      }
-      receiversAndBatches.append((receivers(k), EndOfUpstream()))
-    }
-    receiversAndBatches.toArray
-  }
-
-  override def addTupleToBatch(
-      tuple: ITuple
-  ): Option[(ActorVirtualIdentity, DataPayload)] = {
+  override def selectBatchingIndex(tuple: ITuple): Int = {
     val numBuckets = receivers.length
-    val index = (hashFunc(tuple) % numBuckets + numBuckets) % numBuckets
-    batches(index)(currentSizes(index)) = tuple
-    currentSizes(index) += 1
-    if (currentSizes(index) == batchSize) {
-      currentSizes(index) = 0
-      val retBatch = batches(index)
-      batches(index) = new Array[ITuple](batchSize)
-      return Some((receivers(index), DataFrame(retBatch)))
-    }
-    None
-  }
-
-  override def reset(): Unit = {
-    initializeInternalState(receivers)
-  }
-
-  private[this] def initializeInternalState(_receivers: Array[ActorVirtualIdentity]): Unit = {
-    batches = new Array[Array[ITuple]](_receivers.length)
-    for (i <- _receivers.indices) {
-      batches(i) = new Array[ITuple](batchSize)
-    }
-    currentSizes = new Array[Int](_receivers.length)
+    (hashFunc(tuple) % numBuckets + numBuckets) % numBuckets
   }
 }
