+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/sendsemantics/datatransferpolicy/RoundRobinPolicy.scala
@@ -11,40 +11,11 @@ class RoundRobinPolicy(
     policyTag: LinkIdentity,
     batchSize: Int,
     receivers: Array[ActorVirtualIdentity]
-) extends DataSendingPolicy(policyTag, batchSize, receivers) {
+) extends ParallelBatchingPolicy(policyTag, batchSize, receivers) {
   var roundRobinIndex = 0
-  var batch: Array[ITuple] = new Array[ITuple](batchSize)
-  var currentSize = 0
 
-  override def noMore(): Array[(ActorVirtualIdentity, DataPayload)] = {
-    val ret = new ArrayBuffer[(ActorVirtualIdentity, DataPayload)]
-    if (currentSize > 0) {
-      ret.append((receivers(roundRobinIndex), DataFrame(batch.slice(0, currentSize))))
-    }
-    receivers.foreach { receiver =>
-      ret.append((receiver, EndOfUpstream())) // send end to all receivers
-    }
-    ret.toArray
-  }
-
-  override def addTupleToBatch(
-      tuple: ITuple
-  ): Option[(ActorVirtualIdentity, DataPayload)] = {
-    batch(currentSize) = tuple
-    currentSize += 1
-    if (currentSize == batchSize) {
-      currentSize = 0
-      val retBatch = batch
-      roundRobinIndex = (roundRobinIndex + 1) % receivers.length
-      batch = new Array[ITuple](batchSize)
-      return Some((receivers(roundRobinIndex), DataFrame(retBatch)))
-    }
-    None
-  }
-
-  override def reset(): Unit = {
-    batch = new Array[ITuple](batchSize)
-    roundRobinIndex = 0
-    currentSize = 0
+  override def selectBatchingIndex(tuple: ITuple): Int = {
+    roundRobinIndex = (roundRobinIndex + 1) % receivers.length
+    roundRobinIndex
   }
 }
